refactor(proxy): extract attemptCandidate from ChatCompletions (phase 6)
Move the failover loop body (phase D3–D11) into attemptCandidate, returning
a candidateOutcome (failover / served / fatal). The orchestrator's loop
collapses to the deadline + backoff checks plus a single call:

    if h.attemptCandidate(...) != outcomeFailover { return }

attemptCandidate owns the per-attempt request contexts: failoverCtx is
cancelled via a deferred call and the 400-retry context via a second, so the
~8 manual cancel sites become two defers that fire on every return — after
the streaming/non-streaming dispatch has consumed the body, preserving the
"cancel only after body consumed" ordering. Accumulating state (dial time,
proxy overhead, lastErr, failoverAttempt) moves onto requestState; dialMs is
now attempt-local.

ChatCompletions is now a ~63-line orchestrator (was ~675). No behavior
change: identical breaker calls, failRequest rows, error responses, and
cancel lifecycle. Proxy + failover suite green under -race; lint clean.
